+++ src/components/Navbar.tsx
@@ -2,11 +2,11 @@ import { RiDiscordLine } from "react-icons/ri";
 import { FiLogOut, FiMapPin } from "react-icons/fi";
 import { RxStar } from "react-icons/rx";
 import { useState, useEffect, useMemo } from "react";
-import { useAuth } from "@/providers/hooks";
+import { useAuth } from "@/providers";
 import Hamburger from "hamburger-react";
 import { Link, useLocation } from "react-router-dom";
 import { Logo } from "@/assets";
-import { useRouteDefinitions, useRouter } from "@/providers/routes.provider";
+import { useUser, useRouteDefinitions, useRouter } from "@/providers";
 import {
     CalendarDaysIcon,
     CodeBracketIcon,
@@ -15,8 +15,7 @@ import {
     TicketIcon,
     UserGroupIcon,
 } from "@heroicons/react/24/outline";
-import { useApplications } from "@/providers/applications.provider";
-import { useUser } from "@/providers/auth.provider";
+import { useApplications } from "@/providers/ApplicationsProvider";
 
 export const Navbar = () => {
     const { logout } = useAuth();
