+++ src/navigation/AccessControl/index.ts
@@ -1,6 +1,6 @@
 export { AccessControl } from "./AccessControl";
 export type {
-    AccessControlContext,
-    AccessControlProps,
     AccessControlFn,
-} from "./AccessControl";
+    AccessControlProps,
+    AccessControlContext,
+} from "./AccessControl.types";
