+++ src/pages/Home/Home.page.tsx
@@ -2,7 +2,7 @@ import { GoldenHawk, IpadKidHawks } from "@/assets";
 import { Card, Accordion, SocialIcons, Modal } from "@components";
 import { Button } from "@chakra-ui/react";
 import { faqs, sponsors, importantDateTimes } from "@data";
-import { useAuth } from "@/providers/auth.provider";
+import { useAuth } from "@/providers";
 import { useState } from "react";
 import { toaster } from "@/components/ui/toaster";
 import { withdrawRSVP } from "@/services/firebase/rsvp";
