+++ src/pages/miscellaneous/VerifyRSVP.page.tsx
@@ -1,6 +1,6 @@
 import { LoadingAnimation } from "@/components";
 import { Button } from "@chakra-ui/react";
-import { useAuth } from "@/providers/auth.provider";
+import { useAuth } from "@/providers";
 import { verifyRSVP } from "@/services/firebase/rsvp";
 import { toaster } from "@/components/ui/toaster";
 import { useNavigate } from "react-router-dom";
