+++ src/providers/ApplicationsProvider/ApplicationsProvider.tsx
@@ -1,29 +1,8 @@
-import { createContext, useContext, useCallback } from "react";
-import { useUser } from "./auth.provider";
+import { useCallback } from "react";
+import { useUser } from "@/providers";
 import { getUserApplications } from "@/services/firebase/application";
-import type { ApplicationData } from "@/components/forms/types";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
-
-/**
- * Context value for the applications provider
- */
-interface ApplicationsContextValue {
-    /** List of user's application data */
-    applications: ApplicationData[];
-    /** Function to refresh applications data */
-    refreshApplications: () => void;
-    /** Loading state of applications data */
-    isLoading: boolean;
-}
-
-/**
- * Context for managing application data
- */
-const ApplicationsContext = createContext<ApplicationsContextValue>({
-    applications: [],
-    refreshApplications: async () => {},
-    isLoading: false,
-});
+import { ApplicationsContext } from "./context";
 
 /**
  * Provider component for managing application data
@@ -63,11 +42,4 @@ export const ApplicationsProvider = ({
             {children}
         </ApplicationsContext.Provider>
     );
-};
-
-/**
- * Hook for accessing application data
- */
-export const useApplications = () => {
-    return useContext(ApplicationsContext);
 };
