+++ src/providers/ApplicationsProvider/hooks.ts
@@ -1,0 +1,9 @@
+import { useContext } from "react";
+import { ApplicationsContext } from "@/providers/ApplicationsProvider/context";
+
+/**
+ * Hook for accessing application data
+ */
+export const useApplications = () => {
+    return useContext(ApplicationsContext);
+};
