+++ src/providers/ApplicationsProvider/index.ts
@@ -1,0 +1,2 @@
+export { ApplicationsProvider } from "./ApplicationsProvider";
+export { useApplications } from "./hooks";
