+++ src/providers/AuthProvider/context.ts
@@ -1,0 +1,14 @@
+import { createContext } from "react";
+import { AuthContextValue } from "./types";
+
+export const AuthContext = createContext<AuthContextValue>({
+    currentUser: null,
+    userApp: null,
+    login: async () => {},
+    logout: async () => {},
+    createAccount: async () => {},
+    resetPassword: async () => {},
+    loginWithProvider: async () => {},
+    reloadUser: async () => {},
+    refreshUserApp: async () => {},
+});
