+++ src/providers/AuthProvider/index.ts
@@ -1,0 +1,9 @@
+export { AuthProvider } from "./AuthProvider";
+export { useAuth, useUser } from "./hooks";
+export type {
+    UserType,
+    UserWithClaims,
+    ProviderName,
+    AuthMethod,
+    AuthContextValue,
+} from "./types";
