--- src/providers/hooks.ts
@@ -1,1 +1,0 @@
-export { useAuth } from "./auth.provider";
