+++ src/providers/index.tsx
@@ -1,7 +1,23 @@
-/*
- *
- * @provider AuthProvider
- *
- */
+export { ApplicationsProvider } from "./ApplicationsProvider";
 
-export { AuthProvider } from "./auth.provider";
+export { AuthProvider, useUser, useAuth } from "./AuthProvider";
+export type {
+    UserWithClaims,
+    UserType,
+    ProviderName,
+    AuthMethod,
+    AuthContextValue,
+} from "./AuthProvider";
+
+export {
+    RoutesProvider,
+    useRouter,
+    useHeaderInfo,
+    useRouteDefinitions,
+} from "./RoutesProvider";
+export type {
+    HeaderInfo,
+    RoutesContextValue,
+    PathObject,
+    RouteConfig,
+} from "./RoutesProvider";
