--- src/providers/types.ts
@@ -1,1 +1,0 @@
-export type { ProviderName } from "./auth.provider";
